Workaround by omit using super
Sorry, I kept thinking about this so needed to try it out :sweat-smile:

So, as best as I understand it, what dialyzer complains about here
is this [inside of defexception](https://github.com/elixir-lang/elixir/blob/7b20c281d521aa7aa2ad2baa1e9ae6c579d79d0c/lib/elixir/lib/kernel.ex#L5549C1-L5564C10):

```

      if Map.has_key?(struct, :message) do
        Elixir.Kernel.@(impl(true))

        def message(exception) do
          exception.message
        end

        defoverridable message: 1

        Elixir.Kernel.@(impl(true))

        def exception(msg) when Kernel.is_binary(msg) do
          exception(message: msg)
        end
      end
```

That is we define a clause featuring `is_binary` but we never
call that function with a binary. Another workaround that I
applied was to just have a catch all clause where I call it with
a binary. Not exactly elegant, but it worked.

That said, I also dislike `super` and lived most of my life
blissfully unaware of its existence until some point last year :D

So, this workaround foregoes the `super` call and instead we
kind of reimplement the original `def exception` but it's only
one line so I think that this is fine. I think dialyzer doesn't
see us calling the `super` and hence doesnt complain and hence
this is "fine".